+++ scripts/test-projects.test-support.mts
@@ -2522,7 +2522,10 @@ const SEMANTIC_TOOLING_TARGET_PATTERNS: Array<[RegExp, string[]]> = [
   [/^scripts\/ci-changed-scope\.mjs$/u, [...changedScopeTests, "control-ui-i18n"]],
   [/^scripts\/check-changed\.(?:mjs|mts)$/u, ["changed-lanes"]],
   [/^scripts\/changed-lanes\.(?:mjs|mts)$/u, ["changed-lanes"]],
-  [/^scripts\/(?:lib\/tsx-cli-shim|tsx)\.mjs$/u, ["direct-run-entrypoints", "lint-status"]],
+  [
+    /^scripts\/(?:lib\/tsx-cli-shim|tsx)\.mjs$/u,
+    ["direct-run-entrypoints", "lint-status", "local-check-runtime"],
+  ],
   [
     new RegExp(
       [
